scripts/requirements.txt
globus-sdk
python-dotenv